Finishing responsive, minor component, etc Fixes #116
