+++ src/pages/boardExtended/BoardExtended.js
@@ -8,7 +8,7 @@ import { MailOutlined } from '@ant-design/icons'
 import { WorkspaceContextProvider } from '../../context/WorkspaceContext';
 import { FloatingButton } from "../../components/floatingButton/FloatingButton";
 import ProgressBar from '../../components/progressBar/ProgressBar';
-import BoardOnBoard from '../../components/boardOnBoard/BoardOnBoard';
+import TaskOnBoard from '../../components/taskOnBoard/TaskOnBoard';
 import { BASE_API_URL } from '../../constants/urls';
 import { getToken } from '../../utils/authentication';
 import axios from 'axios';
@@ -73,7 +73,7 @@ const BoardExtended = () => {
                                     </div>
                                     <div className="line-of-boards">
                                         <div className="line-of-boards-item">
-                                            <BoardOnBoard board_id={board_id} board_name={activeBoard.board_name} workspace_name={activeWorkspace.workspace_name} />
+                                            <TaskOnBoard board_id={board_id} board_name={activeBoard.board_name} workspace_name={activeWorkspace.workspace_name} />
                                         </div>
                                     </div>
                                 </div>
