+++ src/pages/dashboard/Dashboard.css
@@ -1,6 +1,16 @@
 .spacer{
 }
 
+.workspace-title-dashboard-1{
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+}
+
+.workspace-title-dashboard-1-2{
+    font-size: 30px;
+}
+
 .main-layout{
     background-color: #F5F8FA;
 }
@@ -10,6 +20,10 @@
     font-weight: bold;
     background-color: #F5F8FA;
     margin-bottom: 20px;
+}
+
+.workspace-title-dashboard{
+    font-size: larger;
 }
 
 .layout-main{
